`useEffect` for the `POST`
any side-effects should happen in a `useEffect` hook. This commit does
that by creating a new hook which does the posting to a server,
resetting of the todo input, and focusing of the `inputEl`. The
[dependencies
array](https://reactjs.org/docs/hooks-reference.html#conditionally-firing-an-effect)
specifies that whenever the `isSaving` state changes, this effect should
execute. That's only half of the story here. We additionally check
_inside_ of the hook that the value of `isSaving` is `true`. Then and
only then do we go ahead with the post. You'll also notice that I'm not
specifying `newTodo` in the dependency array. That's because I don't
need this effect to run every time the value is updated, only when it's
submitted.